requirements.txt
altgraph==0.17.4
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
chardet==5.2.0
cssselect2==0.8.0
dj-database-url==2.3.0
Django==5.1.6
django-chartjs==2.3.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
django-model-utils==5.0.0
django-notifications-hq==1.8.3
djangorestframework==3.15.2
drf-nested-routers==0.94.1
Faker==37.0.0
fonttools==4.58.0
gunicorn==23.0.0
jsonfield==3.1.0
numpy==2.2.6
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyphen==0.17.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32-ctypes==0.2.3
reportlab==4.4.1
setuptools==76.0.0
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.0
tzdata==2025.1
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
